refactor(#452): split the wirenpc composition root into role-scoped files (pure moves)
wirenpc.go was one 1567-line file doing three jobs. Split it by role, moving
every symbol verbatim (identical bodies, package unchanged, no signature or
behavior change):

- wirenpc.go   — package doc, tuning consts, npcSpec/hardcodedNPC, Config
- run.go       — Run / RunFromDB entry points, schema fail-fast, voice-gap log
- reconnect.go — the backed-off reconnect/lifecycle loop (issue #44/#141)
- connect.go   — one connect-and-serve cycle: client acquisition, join, audio
- pipeline.go  — pipeline assembly: buildConversation, engine/stream factories
- mutewiring.go, tapewiring.go, highlightwiring.go — per-feature wiring helpers
- butlertext.go — Butler text delivery (poster + splitter)

The only non-move is buildConversation's doc block, which sat stranded above
wireMutes and now sits on its function.

Part of #452; the buildConversation config struct and the shared Discord
splitter follow in separate commits.
