+++ internal/wirenpc/wirenpc.go
@@ -12,46 +12,22 @@ package wirenpc
 
 import (
 	"context"
-	"database/sql"
-	"fmt"
 	"log/slog"
-	"os"
-	"strconv"
 	"time"
 
-	"github.com/disgoorg/disgo"
 	"github.com/disgoorg/disgo/bot"
-	"github.com/disgoorg/disgo/discord"
-	"github.com/disgoorg/disgo/gateway"
-	"github.com/disgoorg/disgo/rest"
-	"github.com/disgoorg/snowflake/v2"
 	"github.com/google/uuid"
-	"github.com/jackc/pgx/v5/pgxpool"
-	_ "github.com/jackc/pgx/v5/stdlib" // registers the database/sql "pgx" driver for the goose-backed schema check
 
 	"github.com/MrWong99/Glyphoxa/internal/highlight"
-	"github.com/MrWong99/Glyphoxa/internal/llmbuild"
 	"github.com/MrWong99/Glyphoxa/internal/observe"
-	"github.com/MrWong99/Glyphoxa/internal/storage"
-	"github.com/MrWong99/Glyphoxa/internal/storage/crypto"
 	"github.com/MrWong99/Glyphoxa/internal/tape"
 	"github.com/MrWong99/Glyphoxa/pkg/tool"
 	gxvoice "github.com/MrWong99/Glyphoxa/pkg/voice"
-	"github.com/MrWong99/Glyphoxa/pkg/voice/address"
 	"github.com/MrWong99/Glyphoxa/pkg/voice/agent"
-	"github.com/MrWong99/Glyphoxa/pkg/voice/agenttool"
-	"github.com/MrWong99/Glyphoxa/pkg/voice/llm"
 	"github.com/MrWong99/Glyphoxa/pkg/voice/orchestrator"
-	"github.com/MrWong99/Glyphoxa/pkg/voice/retry"
-	"github.com/MrWong99/Glyphoxa/pkg/voice/stt"
-	stteleven "github.com/MrWong99/Glyphoxa/pkg/voice/stt/elevenlabs"
 	"github.com/MrWong99/Glyphoxa/pkg/voice/tts"
 	ttseleven "github.com/MrWong99/Glyphoxa/pkg/voice/tts/elevenlabs"
-	"github.com/MrWong99/Glyphoxa/pkg/voice/vad"
-	"github.com/MrWong99/Glyphoxa/pkg/voice/vad/silero"
 	"github.com/MrWong99/Glyphoxa/pkg/voice/voiceevent"
-	"github.com/MrWong99/Glyphoxa/pkg/voice/wire"
-	"github.com/MrWong99/Glyphoxa/pkg/voice/wire/codec"
 )
 
 const (
@@ -107,118 +83,6 @@ const (
 	// its own turn.
 	floorCoalesceWindow = 600 * time.Millisecond
 )
-
-// reconnectPolicy bounds how the live voice loop backs off between failed or
-// dropped Discord connections (issue #44). A serving voice pod must not
-// crashloop because Discord is briefly unreachable: Run keeps serving /healthz
-// and /readyz (DB-backed) and retries on this schedule instead of exiting.
-// Capped exponential, no jitter — one pod reconnecting to one gateway has no
-// thundering herd to spread.
-type reconnectPolicy struct {
-	initial time.Duration
-	max     time.Duration
-	factor  float64
-	// healthyAfter is how long a cycle must serve after connected() fires before
-	// it counts as a healthy session and resets the backoff to initial. A cycle
-	// that joins but fails sooner (codec-less build, broken ONNX init — issue
-	// #141) is a connect failure: the delay keeps growing to its cap instead of
-	// retrying the Discord voice join at 1 Hz forever. Zero means reset on join.
-	healthyAfter time.Duration
-	// sleep blocks for d or until ctx is cancelled (returns ctx.Err() if
-	// cancelled first). Injected so tests drive the backoff without real waits.
-	sleep func(ctx context.Context, d time.Duration) error
-	// now reports the current time for the healthyAfter measurement. Injected so
-	// tests fake a long-serving session in milliseconds; nil means time.Now.
-	now func() time.Time
-}
-
-// healthySessionDuration is how long a session must serve post-join before the
-// reconnect backoff forgives past failures and resets to the initial delay.
-// Sized to the backoff cap: a session must outlive the maximum backoff before
-// the loop trusts it, so a persistent join-then-fail cycle (issue #141) settles
-// at cap cadence instead of resetting to 1 Hz.
-const healthySessionDuration = 30 * time.Second
-
-func defaultReconnectPolicy() reconnectPolicy {
-	return reconnectPolicy{
-		initial:      time.Second,
-		max:          30 * time.Second,
-		factor:       2,
-		healthyAfter: healthySessionDuration,
-		sleep:        sleepCtx,
-		now:          time.Now,
-	}
-}
-
-// sleepCtx blocks for d or until ctx is cancelled, returning ctx.Err() on
-// cancel. A timer (not time.Sleep) so a cancelled ctx returns immediately.
-func sleepCtx(ctx context.Context, d time.Duration) error {
-	t := time.NewTimer(d)
-	defer t.Stop()
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	case <-t.C:
-		return nil
-	}
-}
-
-func nextDelay(d time.Duration, p reconnectPolicy) time.Duration {
-	next := time.Duration(float64(d) * p.factor)
-	if next > p.max {
-		return p.max
-	}
-	return next
-}
-
-// runWithReconnect calls attempt repeatedly, keeping the process alive across
-// failed or dropped Discord connections. It returns nil (clean shutdown) ONLY
-// when ctx is cancelled; every other return from attempt — an error OR a clean
-// session-close (nil) — is a lost connection and triggers a backed-off
-// reconnect. attempt is handed a connected callback it calls once the join
-// succeeds; the backoff resets to initial only if the cycle then serves for at
-// least p.healthyAfter (issue #141), so a long-lived session that later drops
-// reconnects promptly while a join-then-immediate-fail cycle keeps growing its
-// delay instead of hammering the Discord voice join at 1 Hz.
-func runWithReconnect(ctx context.Context, log *slog.Logger, p reconnectPolicy, attempt func(ctx context.Context, connected func()) error) error {
-	now := p.now
-	if now == nil {
-		now = time.Now
-	}
-	delay := p.initial
-	for {
-		// connectedAt is written by the callback inside attempt and read after
-		// attempt returns — connectAndServe invokes it synchronously, same
-		// goroutine, so no lock. The delay is only consumed post-return, so
-		// deciding the reset here is equivalent to arming a timer on connect.
-		var connectedAt time.Time
-		err := attempt(ctx, func() { connectedAt = now() })
-		if ctx.Err() != nil {
-			return nil // shutdown requested — stop retrying, exit clean (fixes SIGTERM->exit1)
-		}
-		// A fatal, non-retryable gateway rejection (bad Bot token, disallowed
-		// intents, gateway reject) can never succeed on retry, so stop at once and
-		// surface the classification instead of backing off forever (#123). The
-		// session Manager reads this to record the persisted status as failed. A
-		// transient failure (or a clean nil return) falls through to the backoff.
-		if fe := classifyFatal(err); fe != nil {
-			log.Error("voice connection failed fatally; not retrying", "err", fe, "reason", fe.Reason)
-			return fe
-		}
-		if !connectedAt.IsZero() && now().Sub(connectedAt) >= p.healthyAfter {
-			delay = p.initial // served healthily — forgive past failures (issue #44)
-		}
-		if err != nil {
-			log.Warn("voice connection failed; reconnecting", "err", err, "backoff", delay)
-		} else {
-			log.Info("voice session ended; reconnecting", "backoff", delay)
-		}
-		if serr := p.sleep(ctx, delay); serr != nil {
-			return nil // ctx cancelled during backoff — clean shutdown
-		}
-		delay = nextDelay(delay, p)
-	}
-}
 
 // BartPersona is the Character NPC Persona (CONTEXT.md "Persona") for the MVP
 // slice. Exported so the `seed` command writes the same Persona text the in-code
@@ -462,544 +326,6 @@ type Config struct {
 	GMSpeaker func(speakerID string) bool
 }
 
-// RunFromDB loads the bound Active Campaign's Character NPCs from Postgres (via
-// the task-#8 storage layer, scoped by cfg.CampaignID — #323) and runs the live
-// voice loop with them, instead of the in-code NPC. The campaign id is the
-// selection the Voice Session carries; an empty id fails loudly rather than
-// silently voicing the seed roster. pool is an already-open pgxpool the caller
-// owns (and closes) — voice mode
-// opens exactly one pool that ALSO backs the /readyz probe, and all/web mode
-// hands in its existing request pool, so the voice path never opens a second
-// duplicate handle. This is the task-#5 DB-load path: the only thing it changes
-// versus [Run] is the *source* of the NPC's Persona/Voice/identity — the
-// assembled pipeline is identical.
-//
-// cipher decrypts the saved BYOK provider credentials (issue #69, ADR-0004): a
-// real saved key (last4 != "env") drives the session decrypted, while the seeded
-// "env" placeholder falls back to the adapter's own env var (the hybrid policy,
-// ADR-0039). A nil cipher is fine when every config is the env placeholder — the
-// no-$GLYPHOXA_SECRET self-host path — but a real saved key with no cipher is a
-// clear startup error, never a silent fall back to ENV.
-func RunFromDB(ctx context.Context, cfg Config, pool *pgxpool.Pool, cipher *crypto.Cipher) error {
-	log := cfg.Logger
-	if log == nil {
-		log = slog.New(slog.DiscardHandler)
-	}
-
-	// Fail fast on a stale schema BEFORE any other DB interaction (ADR-0031):
-	// serving Modes (voice) never auto-migrate, so a DB behind the embedded
-	// migrations must refuse to start with the actionable `migrate up` message
-	// rather than running queries against a schema the code no longer matches.
-	// This runs before loadCampaignNPCs (the first query). The schema check needs a
-	// database/sql handle (goose's API), which the pgxpool can't provide, so the
-	// dsn is recovered from the pool's own config — no second connection string
-	// threaded through the callers.
-	if err := ensureSchemaCurrent(ctx, pool.Config().ConnString()); err != nil {
-		return err
-	}
-
-	st := storage.New(pool)
-	npcs, primary, campaign, err := loadCampaignNPCs(ctx, st, cfg.CampaignID)
-	if err != nil {
-		return err
-	}
-	for _, npc := range npcs {
-		log.Info("loaded NPC from DB", "npc", npc.name, "agentID", npc.agentID)
-	}
-	// Surface the #224 failure mode at session start, loudly: an NPC that hydrated
-	// with an empty VoiceID will be silent at synthesis time (elevenlabs.Synthesize
-	// rejects an empty Voice.VoiceID). Logging it here — once, at load — beats the
-	// per-turn WARN storm that hid the cause during live validation.
-	logVoiceGaps(log, npcs)
-
-	// Resolve the session's BYOK keys from the saved provider_config (issue #69).
-	// A decryption failure (e.g. a real saved key with the wrong/absent cipher)
-	// is fatal here, before any Discord connection — the operator sees a clear
-	// error instead of an NPC that silently ran on the wrong (env) key.
-	keys, err := resolveSessionKeys(ctx, st, campaign.TenantID, primary, cipher)
-	if err != nil {
-		return err
-	}
-
-	cfg.npcs = npcs
-	cfg.keys = keys
-	cfg.llmProviderID = llmProviderID(primary.LLMConfig)
-	cfg.language = campaign.Language
-
-	// Rollover tape (#306, ADR-0051): armed ONLY when the Campaign opted in
-	// (tape_armed). Seed it with the individually-consenting Speakers; agent speech
-	// is always captured regardless. The tape lives across reconnect cycles for the
-	// whole session and is discarded here at session end — only promoted Highlights
-	// outlive it (a later slice). Default OFF: an unarmed campaign gets no tape, no
-	// taps, no capture whatsoever.
-	if campaign.TapeArmed {
-		consented, err := st.ListTapeConsent(ctx, cfg.CampaignID)
-		if err != nil {
-			return fmt.Errorf("wirenpc: load tape consent: %w", err)
-		}
-		tp := tape.New(tape.Window, consented, log)
-		defer tp.Close()
-		cfg.Tape = tp
-		cfg.TapeConsent = st
-		log.Info("rollover tape armed", "campaign", cfg.CampaignID, "consented_speakers", len(consented))
-	}
-
-	return Run(ctx, cfg)
-}
-
-// logVoiceGaps emits an ERROR for every loaded NPC whose Voice has an empty
-// VoiceID (#224, AC6): such an NPC is unsynthesizable and will be silent, so the
-// gap is surfaced ONCE at session start instead of failing per-turn at synthesis
-// time. It is unconditional on the DB path — the check is the empty VoiceID
-// itself, not the NPC's addressability or TTS binding, because an empty VoiceID
-// is always the silent-NPC condition.
-func logVoiceGaps(log *slog.Logger, npcs []npcSpec) {
-	for _, npc := range npcs {
-		if npc.voice.VoiceID != "" {
-			continue
-		}
-		// The Butler is a legitimately text-only Agent (#299): a voiceless Butler
-		// answers via its TextSink into the channel chat, so an empty VoiceID is
-		// expected, not the silent-NPC failure. Surface it at INFO, not ERROR.
-		if npc.role == voiceevent.AgentRoleButler {
-			log.Info("Butler has no voice configured; it will answer as text only",
-				"npc", npc.name, "agentID", npc.agentID)
-			continue
-		}
-		log.Error("NPC has no synthesizable voice (empty VoiceID); it will be silent",
-			"npc", npc.name, "agentID", npc.agentID)
-	}
-}
-
-// ensureSchemaCurrent verifies the DB at dsn is migrated to the latest embedded
-// schema version, returning the storage layer's actionable version-mismatch
-// error (verbatim) if it is behind. This is the ADR-0031 fail-fast guard for
-// serving Modes: [RunFromDB] calls it once at startup, after the pool opens and
-// before any other query, so a process can never serve against a stale schema.
-//
-// [storage.EnsureCurrent] needs a database/sql handle on the pgx stdlib driver
-// (goose's API; the app's own queries use the pgxpool). That handle exists only
-// for this check, so it is opened from the same dsn and closed immediately —
-// keeping the seam free of the live voice loop and Discord, so it is testable on
-// its own against a real Postgres.
-func ensureSchemaCurrent(ctx context.Context, dsn string) error {
-	db, err := sql.Open("pgx", dsn)
-	if err != nil {
-		return fmt.Errorf("wirenpc: open schema-check handle: %w", err)
-	}
-	defer db.Close()
-	return storage.EnsureCurrent(ctx, db)
-}
-
-// EnsureSchemaCurrent is the exported ADR-0031 fail-fast guard for callers that
-// must run a DB query BEFORE [RunFromDB] does its own internal check. The
-// standalone voice entrypoint resolves the Active Campaign before RunFromDB
-// (#323), so it runs this first to keep the stale-schema refusal (the actionable
-// `migrate up` message) ahead of any other query. It delegates to the same
-// self-contained schema-check seam RunFromDB uses.
-func EnsureSchemaCurrent(ctx context.Context, dsn string) error {
-	return ensureSchemaCurrent(ctx, dsn)
-}
-
-// Run builds and runs the live NPC voice loop until ctx is cancelled. It joins
-// the configured voice channel, wires the orchestrator pipeline with the
-// production Agent loop, and pumps audio through [wire.Pipeline] in both
-// directions: inbound Opus → DecodeInbound → VAD/STT (hear), and synthesized TTS
-// → tee → serial playback → Opus → Session.Play (speak).
-//
-// Audio requires the real Opus↔PCM [codec]; it is compiled in only under
-// -tags opus (system libopus). A default build links the codec stub, so Run
-// still connects and constructs the whole pipeline but the audio loop fails fast
-// with [wire.ErrCodecUnavailable] on the first inbound frame — the binary is
-// runnable and the wiring complete without the native dependency. Build with
-// -tags "opus dave nolibopusfile" for a hearing, speaking, encrypted NPC.
-func Run(ctx context.Context, cfg Config) error {
-	if len(cfg.npcs) == 0 {
-		cfg.npcs = []npcSpec{hardcodedNPC()}
-	}
-
-	log := cfg.Logger
-	if log == nil {
-		log = slog.New(slog.DiscardHandler)
-	}
-
-	// Config validation is fatal: a bad guild/channel ID can never succeed, so
-	// retrying would crashloop slowly. Parse before the reconnect loop so only
-	// genuinely transient connection failures are retried.
-	guild, err := snowflake.Parse(cfg.Guild)
-	if err != nil {
-		return fmt.Errorf("wirenpc: parse guild ID %q: %w", cfg.Guild, err)
-	}
-	channel, err := snowflake.Parse(cfg.Channel)
-	if err != nil {
-		return fmt.Errorf("wirenpc: parse channel ID %q: %w", cfg.Channel, err)
-	}
-
-	// Keep serving across a briefly unreachable or dropped Discord instead of
-	// exiting (issue #44): cmd/glyphoxa's metrics server (which carries /healthz
-	// and the DB-backed /readyz) lives for ctx independently of this loop, so a
-	// reconnecting voice loop lets the Deployment reach Available without live
-	// Discord creds. Each cycle is one connectAndServe; runWithReconnect backs
-	// off between cycles and returns clean only when ctx is cancelled.
-	//
-	// Note: disgo runs its own bounded reconnect during OpenGateway, so this
-	// policy governs the inter-cycle gap and post-join drops (a session that joins
-	// then later disconnects), not the initial dial retries disgo already handles.
-	runErr := runWithReconnect(ctx, log, defaultReconnectPolicy(),
-		func(ctx context.Context, connected func()) error {
-			return connectAndServe(ctx, cfg, guild, channel, log, connected)
-		})
-	// A non-nil return is a FATAL, terminal failure (a classified *FatalError):
-	// publish the terminal connection.state{failed} with its readable reason so the
-	// Session screen flips to failed without a reload (#123). nil is a clean
-	// ctx-cancel shutdown — no failed frame. Nil-guarded: the env-only/bench paths
-	// carry no shared bus and this is a no-op there.
-	if runErr != nil {
-		publishConnectionState(cfg.Bus, voiceevent.ConnectionFailed, runErr.Error())
-	}
-	return runErr
-}
-
-// publishConnectionState publishes a [voiceevent.ConnectionStateChanged] onto bus,
-// unless bus is nil (the env-only/bench paths carry no shared bus). It is the one
-// seam the connection-state transitions go through so the nil-guard lives in one
-// place (#123).
-func publishConnectionState(bus *voiceevent.Bus, state voiceevent.ConnectionState, detail string) {
-	if bus == nil {
-		return
-	}
-	bus.Publish(voiceevent.ConnectionStateChanged{At: time.Now(), State: state, Detail: detail})
-}
-
-// connectAndServe runs ONE connect-and-serve cycle: build the Discord client,
-// open the gateway, join the channel, assemble the pipeline, and pump audio
-// until ctx is cancelled or the connection drops. It calls connected() once the
-// join succeeds, marking when serving began; the caller resets its backoff only
-// if the cycle then survives the healthy threshold (issue #141), so a
-// long-lived session that later drops reconnects promptly while a
-// join-then-immediate-fail cycle keeps backing off. Any error — or a clean return (a dropped
-// gateway often reports none) — flows back to runWithReconnect, which decides
-// whether to retry; only a cancelled ctx ends the loop.
-// newDiscordClient is the disgo constructor seam (mirrors newLLM/newSTT/newTTS):
-// a test swaps it to assert the owned path calls it and the shared-client path
-// does NOT. Production always uses disgo.New.
-var newDiscordClient = disgo.New
-
-// acquireClient yields the Discord client for one connect-and-serve cycle. When
-// cfg.Client is set it BORROWS the standing shared client (already gateway-open,
-// owned by the presence) and reports owned=false so the caller never closes it;
-// a provider error fails the cycle so runWithReconnect retries. Otherwise it
-// builds and opens a per-cycle client (today's behavior) and reports owned=true.
-func acquireClient(ctx context.Context, cfg Config, bus *voiceevent.Bus, log *slog.Logger) (client *bot.Client, owned bool, err error) {
-	if cfg.Client != nil {
-		c, err := cfg.Client(ctx)
-		if err != nil {
-			return nil, false, fmt.Errorf("wirenpc: standing Discord client unavailable: %w", err)
-		}
-		return c, false, nil
-	}
-
-	// Per-cycle client: DAVE/MLS is wired at construction (it cannot be enabled
-	// after disgo builds its VoiceManager). DaveOption() is a no-op stub unless the
-	// binary was built with -tags dave; NewManager(WithDave(true)) then warns if
-	// encryption was expected but unavailable.
-	opts := []bot.ConfigOpt{
-		// Own disgo's logger explicitly (A1): route it through the same filtered
-		// app logger so the benign DAVE-decrypt noise is tamed even if disgo ever
-		// stops reading slog.Default().
-		bot.WithLogger(log),
-		bot.WithDefaultGateway(),
-		// disgo's default intents are IntentsNone, so the bot never receives its
-		// own VoiceStateUpdate — leaving the voice conn's ChannelID nil and
-		// segfaulting disgo's voice gateway on the VoiceServerUpdate join path.
-		// GuildVoiceStates (+Guilds) is the minimum to populate that state.
-		bot.WithGatewayConfigOpts(gateway.WithIntents(
-			gateway.IntentGuilds | gateway.IntentGuildVoiceStates,
-		)),
-		gxvoice.DaveOption(),
-	}
-	// Standalone voice mode has no boot-owned presence to answer the tape consent
-	// disclosure's buttons (#306, finding 5), so THIS per-cycle client must carry
-	// the listener itself — otherwise every Consent/Revoke press fails. The all-mode
-	// shared client (cfg.Client != nil, handled above) gets it from the presence.
-	//
-	// It MUST run off the gateway read goroutine: the handler does a DB upsert, a
-	// publish, an authoritative reconcile (2nd DB round-trip + tape ctrl) and a
-	// Discord REST reply, and stalling the read goroutine on all that misses
-	// heartbeats → reconnect churn (the exact ADR-0010 hazard the presence guards
-	// with async events — see internal/presence). So enable async event delivery
-	// alongside the listener, mirroring defaultClientBuilder. Gated on the listener
-	// so a nil-Tape cycle keeps today's synchronous, listener-free client unchanged.
-	if l := tapeConsentListener(cfg.TapeConsent, bus, log); l != nil {
-		opts = append(opts,
-			bot.WithEventListenerFunc(l),
-			bot.WithEventManagerConfigOpts(bot.WithAsyncEventsEnabled()),
-		)
-	}
-	c, err := newDiscordClient(cfg.Token, opts...)
-	if err != nil {
-		return nil, false, fmt.Errorf("wirenpc: build Discord client: %w", err)
-	}
-	if err := c.OpenGateway(ctx); err != nil {
-		c.Close(context.Background())
-		return nil, false, fmt.Errorf("wirenpc: open gateway: %w", err)
-	}
-	return c, true, nil
-}
-
-func connectAndServe(ctx context.Context, cfg Config, guild, channel snowflake.ID, log *slog.Logger, connected func()) error {
-	// Per-cycle context: everything this cycle spawns — the stage subscriber's
-	// TTL-sweep goroutine (stageSub.Start), the Discord gateway, and the audio
-	// loop — is bound to cycleCtx, so the deferred cancel reaps it at cycle end.
-	// Without this a flapping Discord (issue #44) would leak one sweeper goroutine
-	// per reconnect: the outer ctx only ends at process shutdown. cycleCtx is a
-	// child of ctx, so a cancelled process still unwinds promptly (pipe.Run et al.
-	// observe the cancellation), and runWithReconnect checks the OUTER ctx to
-	// decide shutdown vs reconnect.
-	cycleCtx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	// The orchestrator bus is created here (not deeper) so the connection-state
-	// transitions ride the SAME bus as the pipeline's events (#123). The web tier
-	// injects ONE process-wide bus (cfg.Bus, issue #73) so the SSE relay subscribes
-	// once and keeps observing events across reconnect cycles and sessions; the
-	// env-only voice/bench paths leave it nil and get a fresh per-cycle bus (today's
-	// behavior, unchanged). Hoisted above acquireClient so {connecting} is published
-	// first thing each cycle, BEFORE the (possibly fatal) Discord client is acquired.
-	bus := cfg.Bus
-	if bus == nil {
-		bus = voiceevent.NewBus()
-	}
-	publishConnectionState(bus, voiceevent.ConnectionConnecting, "")
-
-	// Discord client: either the standing shared client the presence owns
-	// (cfg.Client, #102) or a per-cycle client this loop builds and closes. The
-	// shared client is already gateway-open and must NOT be closed here.
-	client, owned, err := acquireClient(cycleCtx, cfg, bus, log)
-	if err != nil {
-		return err
-	}
-	if owned {
-		defer client.Close(context.Background())
-	}
-
-	mgr := gxvoice.NewManager(client,
-		gxvoice.WithLogger(log),
-		gxvoice.WithDave(gxvoice.DaveAvailable()),
-		gxvoice.WithMetrics(cfg.Metrics),
-	)
-	defer mgr.Close()
-
-	sess, err := mgr.Open(cycleCtx, guild, channel)
-	if err != nil {
-		return fmt.Errorf("wirenpc: join voice channel: %w", err)
-	}
-	defer sess.Close()
-	// The join succeeded: mark when serving began. runWithReconnect resets the
-	// backoff only if this cycle survives the healthy threshold (issue #141), so
-	// a session that serves for a while and later drops reconnects on the initial
-	// delay (issue #44) while a join-then-immediate-fail keeps backing off.
-	connected()
-	// Announce the Bot is connected and serving so the Session screen flips
-	// connecting → connected live (#123). Rides the same bus as {connecting}.
-	publishConnectionState(bus, voiceevent.ConnectionConnected, "")
-	log.Info("joined voice channel", "guild", guild, "channel", channel, "npcs", npcNames(cfg.npcs))
-
-	// One Codec instance serves both directions: DecodeInbound (called from the
-	// single Pipeline.Run goroutine) and PlaybackSource (called from the playback
-	// worker) — the codec documents this split as concurrency-safe. codec.New()
-	// is the real Opus transcoder under -tags opus and a fail-fast stub
-	// (ErrCodecUnavailable) otherwise, so this binary needs no build-tag
-	// knowledge: a default build still constructs and runs, just deaf+mute.
-	// Living in the shared Run core, this audio path covers BOTH the hardcoded
-	// and the RunFromDB paths (RunFromDB resolves the NPC then delegates here).
-	//
-	// codec_decode / codec_encode (#125): the codec stamps the per-frame Opus<->PCM
-	// costs against cfg.StageMetrics. A nil StageMetrics (env-only) keeps the no-op
-	// default via WithMetrics; the stub build accepts and ignores the same option.
-	cdc := codec.New(codec.WithMetrics(cfg.StageMetrics))
-
-	// Outbound (speak): the PlaybackPump drives the codec's PlaybackSource onto
-	// the Session, one sentence at a time (Session.Play auto-interrupts, so
-	// overlapping playback would clip sentences). The TeeSynthesizer wraps the
-	// real ElevenLabs synthesizer and tees each synthesized chunk to the pump
-	// while the orchestrator's TTS stage keeps draining-and-dropping it (ADR-0021
-	// intact).
-	//
-	// pump.Close() blocks until the playback worker has exited; the deferred
-	// Close must run BEFORE sess.Close() so a mid-flight Play cannot race the
-	// Session teardown. defers run LIFO and this registers after sess.Close()
-	// (line above), and pipe.Run's own deferred cancel stops the Conversation
-	// first — so teardown order is conv-stop → pump.Close() → sess.Close(), which
-	// is the deterministic ordering the pump's Close() contract requires.
-	// The orchestrator bus was created at the top of this cycle (so the tee can
-	// publish FirstAudio and the pump FirstOpus onto the same bus the conversation's
-	// stages publish on and the metrics/SSE subscribers read) and already carries
-	// this cycle's connection.state{connecting}/{connected} (#123).
-	// tapePumpOptions adds the outbound (agent-speech) tape tap when the campaign is
-	// armed (#306); nil tape → no option → unchanged playback.
-	pump := wire.NewPlaybackPump(sess, cdc, log, bus, tapePumpOptions(cfg.Tape)...)
-	defer pump.Close()
-
-	// cfg.keys.tts is the resolved BYOK TTS key (issue #69): the decrypted saved
-	// key when one is configured, or "" to fall back to ELEVENLABS_API_KEY.
-	teeSynth := wire.NewTeeSynthesizer(newTTS(cfg.keys.tts), pump, bus)
-
-	// Attach the orchestrator-sibling latency subscriber (A2/#10): it derives the
-	// SLO histograms (response_latency, address_detect, per-sentence tts_ttfb) from
-	// the turn-correlated bus events and feeds cfg.StageMetrics. Subscribe wires
-	// the handlers (deferred unsubscribe); Start runs the TTL sweep for the run's
-	// lifetime so abandoned/barged turns don't leak per-turn state. A nil
-	// StageMetrics (keyless) makes the subscriber a no-op via observe.Discard.
-	stageSub := observe.NewStageSubscriber(cfg.StageMetrics, observe.WithTurnLog(log))
-	defer stageSub.Subscribe(bus)()
-	stageSub.Start(cycleCtx)
-
-	// Butler text delivery (#299, #297 decision 2): a poster over the BORROWED
-	// Discord client that writes into the voice channel's text chat, so a Butler
-	// answering as text posts there. It reuses this cycle's client and channel; the
-	// TextSink is set on butler-role specs only inside rosterDepsForLive.
-	textPoster := newVoiceChannelPoster(client, channel)
-
-	conv, roster, cleanup, err := buildConversation(bus, log, cfg.npcs, cfg.language, teeSynth, cfg.StageMetrics, cfg.keys, cfg.llmProviderID, cfg.STTStreaming, cfg.Memory, cfg.Facts, cfg.Mutes, cfg.Gate, cfg.GMSpeaker, cfg.ToolDeps, textPoster, pump, cfg.ClipReplayLoader, orchestrator.ClipSink(pump.HandleSentence))
-	if err != nil {
-		return fmt.Errorf("wirenpc: build pipeline: %w", err)
-	}
-	// cleanup closes the per-cycle VAD session (not the shared Silero engine — see
-	// buildConversation). Without it each reconnect cycle (issue #44) would leak a
-	// Silero session that nothing ever closed.
-	defer cleanup()
-
-	// Mute control (#211): subscribe roster.SetMuted to MuteChanged so a GM mute
-	// (web panel or /glyphoxa mute) de-routes the NPC live, and seed the current
-	// mute state so a mid-session Discord RECONNECT re-applies the mutes onto the
-	// freshly-rebuilt roster. nil Mutes = feature off, so both are inert (voice
-	// standalone / bench unchanged).
-	defer wireMutes(bus, roster, cfg.Mutes)()
-
-	// Rollover-tape consent (#306, ADR-0051): reseed the tape from the durable
-	// consent rows at cycle start and reconcile on every (campaign-filtered)
-	// TapeConsentChanged, and post the in-channel consent disclosure with
-	// grant/revoke buttons. A nil Tape (campaign not armed) makes both inert, so the
-	// loop is unchanged.
-	defer wireTapeConsent(cycleCtx, bus, cfg.Tape, cfg.CampaignID, cfg.TapeConsent, log)()
-	if cfg.Tape != nil {
-		if err := postTapeDisclosure(cycleCtx, client, channel, cfg.CampaignID); err != nil {
-			// A failed disclosure post must not tear down the session — capture is
-			// still consent-gated (only prior consenters are taped), and the operator
-			// can re-post. Log and continue.
-			log.Warn("post tape consent disclosure", "err", err)
-		}
-	}
-
-	// Inbound (hear): the pipeline pumps Session.Inbound through the same Codec's
-	// DecodeInbound into the orchestrator. It tags its inbound counters (A2) with
-	// the guild and shares the run's MetricsRecorder.
-	//
-	// WithSilenceClock drives VAD endpointing during the inbound packet gap a
-	// paused speaker leaves (issue #91): Discord sends a few Opus silence frames
-	// then stops sending entirely, so without a steady silence clock the VAD never
-	// sees the trailing silence and each line lands one utterance late. The clock
-	// runs at the VAD frame geometry (vadSampleRate/vadFrameMs) so silero endpoints
-	// ~vadMinSilenceFrames*vadFrameMs (= 384 ms) after the speaker stops — the
-	// natural cadence the bargeConfirm/floorCoalesce windows already account for.
-	// Session Highlights detector (#307, ADR-0051): built ONLY when the campaign is
-	// armed (Tape set) AND a highlight sink is wired. It subscribes to STTFinal off
-	// the bus and consumes the decoded-PCM tap; both are non-blocking (ADR-0020). A
-	// nil detector (highlights off, or provider build failed) adds no option and no
-	// bus subscriber, so the loop is byte-identical. Per-cycle state: Close on exit
-	// (a leak is a #44-class bug).
-	detector := buildHighlightDetector(cfg, bus, log)
-	if detector != nil {
-		defer detector.Close()
-	}
-
-	// tapeInboundOptions adds the inbound (consented Speaker) tape tap when armed
-	// (#306); nil tape → no option → byte-identical inbound loop. highlightPCMOptions
-	// adds the detector's decoded-PCM tap (#307) when the detector is built.
-	pipeOpts := append([]wire.Option{wire.WithSilenceClock(vadSampleRate, vadFrameMs)}, tapeInboundOptions(cfg.Tape)...)
-	pipeOpts = append(pipeOpts, highlightPCMOptions(detector)...)
-	pipe := wire.NewPipeline(conv, cdc, log, cfg.Guild, cfg.Metrics, pipeOpts...)
-	return pipe.Run(cycleCtx, sess)
-}
-
-// npcMatcher builds the Address Detection matcher for one NPC. With a single
-// non-Address-Only Character NPC and no addressable Butler it uses the scoring
-// Matcher (ADR-0024): the NPC gets a name/alias match AND the single-NPC
-// fallback, so both a named utterance ("Bart, …") and an unnamed one route to it.
-// The whole-word matcher is deliberately not used: it requires a Butler as its
-// unconditional fallback, which this slice does not have, and would leave the NPC
-// silent on every unnamed utterance.
-//
-// Multi-NPC wiring goes through [Roster.AddNPC] (the matcher's roster grows past
-// the first agent via [address.Matcher.Add]); this single-agent constructor is
-// retained as the unit-test seam for the one-NPC routing invariant.
-func npcMatcher(npc npcSpec) *address.Matcher {
-	return address.NewMatcher(address.Config{Language: "en"}, matcherAgent(npc))
-}
-
-// discordMessageLimit is Discord's per-message character cap (#299): a Butler
-// text answer longer than this is split across several ordered messages.
-const discordMessageLimit = 2000
-
-// newVoiceChannelPoster builds the Butler's text-delivery sink (#299, #297
-// decision 2): a func that posts text into the voice channel's text chat over the
-// borrowed Discord client, splitting answers longer than [discordMessageLimit]
-// runes into ordered messages. The ctx bounds each REST call (a barge cancels a
-// mid-post). It is wired as the [agent.Config.TextSink] on butler-role specs only.
-func newVoiceChannelPoster(client *bot.Client, channel snowflake.ID) func(ctx context.Context, text string) error {
-	return func(ctx context.Context, text string) error {
-		for _, part := range splitDiscordMessage(text, discordMessageLimit) {
-			if _, err := client.Rest.CreateMessage(channel, discord.MessageCreate{Content: part}, rest.WithCtx(ctx)); err != nil {
-				return fmt.Errorf("wirenpc: post Butler text to channel %s: %w", channel, err)
-			}
-		}
-		return nil
-	}
-}
-
-// splitDiscordMessage breaks text into ordered chunks each at most limit RUNES
-// (never bytes — a German Butler answer), preferring a newline then a space break
-// so a chunk ends at a natural boundary; a single unbroken run longer than limit
-// is hard-cut. Every rune is delivered (never truncated). It mirrors the recap
-// splitter's contract (internal/presence.splitFollowups); it is duplicated rather
-// than imported because internal/presence imports wirenpc (an import cycle).
-func splitDiscordMessage(text string, limit int) []string {
-	runes := []rune(text)
-	if limit <= 0 || len(runes) <= limit {
-		return []string{text}
-	}
-	var parts []string
-	for len(runes) > limit {
-		cut := limit
-		if i := lastBreakRune(runes, limit, '\n'); i > 0 {
-			cut = i
-		} else if i := lastBreakRune(runes, limit, ' '); i > 0 {
-			cut = i
-		}
-		parts = append(parts, string(runes[:cut]))
-		rest := runes[cut:]
-		if cut < limit && len(rest) > 0 && (rest[0] == '\n' || rest[0] == ' ') {
-			rest = rest[1:] // drop the boundary whitespace we broke on
-		}
-		runes = rest
-	}
-	if len(runes) > 0 {
-		parts = append(parts, string(runes))
-	}
-	return parts
-}
-
-// lastBreakRune returns the last index in [1, limit) where runes[i]==ch, or -1.
-func lastBreakRune(runes []rune, limit int, ch rune) int {
-	for i := limit - 1; i > 0; i-- {
-		if runes[i] == ch {
-			return i
-		}
-	}
-	return -1
-}
-
 // npcNames returns the NPCs' display names for a log line.
 func npcNames(npcs []npcSpec) []string {
 	names := make([]string, len(npcs))
@@ -1026,542 +352,4 @@ func npcVoice() tts.Voice {
 	v := ttseleven.DefaultVoice(elevenGeorgeVoiceID, "en")
 	v.Name = "Bart"
 	return v
-}
-
-// Provider-adapter constructors, injected as package vars so a test can spy on
-// the apiKey each component receives (issue #69). The adapters expose no key
-// getter, so this is the seam that pins the resolved BYOK key reaching its OWN
-// adapter — a slot swap (e.g. groq.New(keys.stt)) or a dropped `cfg.keys = keys`
-// would otherwise revert the feature to ENV while every providerKeys{} test
-// stayed green. Production always uses the real constructors.
-var (
-	newLLM = llmbuild.New
-	newSTT = stteleven.New
-	newTTS = ttseleven.New
-	// newAddressDetector is the constructor seam for the address detector (#280):
-	// buildConversation dispatches through it so a test can capture the detector —
-	// and the Butler GM-gate option threaded onto it — without standing up the
-	// whole Conversation. The live path is orchestrator.NewAddressDetector verbatim.
-	newAddressDetector = orchestrator.NewAddressDetector
-)
-
-// llmProviderID reports the provider id of an LLM Provider Config for [newLLM]
-// dispatch (#272). A nil config — the env-only [Run] path, or a DB Agent with no
-// bound LLM config — yields "", which [llmbuild.New] resolves to the Groq default
-// (ADR-0036), keeping the pre-#272 hardwired-groq behaviour byte-identical.
-func llmProviderID(cfg *storage.ProviderConfig) string {
-	if cfg == nil {
-		return ""
-	}
-	return cfg.Provider
-}
-
-// llmProviderLabel maps an LLM provider id to its bounded [observe.Provider] metric
-// label (#272): the empty id (env-only / nil config) is Groq (ADR-0036). The wired
-// ids equal their observe constants, so the cast is exact. It keeps the per-round
-// LLM spans AND the spend price lookup keyed to the ACTUAL provider — a non-Groq
-// adapter mislabelled groq would price (groq, claude-model) as a miss and trip caps
-// on the frontier default.
-func llmProviderLabel(providerID string) observe.Provider {
-	if providerID == "" {
-		return observe.ProviderGroq
-	}
-	return observe.Provider(providerID)
-}
-
-// buildConversation assembles the orchestrator reactive pipeline: VAD (Silero)
-// → STT (ElevenLabs) → Address Detection → production Reply (the Agent loop over
-// Groq, with the dice Tool granted via the tool-use loop) → TTS (synth).
-//
-// keys are the resolved BYOK provider keys (issue #69, hybrid policy ADR-0039):
-// each adapter is constructed with its component's key, which OVERRIDES that
-// adapter's *_API_KEY env var — except an empty key (the env placeholder, or the
-// env-only [Run] path) keeps today's behavior, where the adapter reads its env
-// var at request time (ADR-0004). So a saved key drives the session and an
-// unconfigured component falls back to ENV.
-//
-// npcs supplies the INITIAL Character NPCs the loop voices — their addressable
-// identity, Persona, and Voice (from the in-code seed or, via [RunFromDB], the
-// database). It assembles them into a [Roster] (one address Matcher + one Cast):
-// the detector routes against the Matcher and the reply stream multiplexes across
-// the Cast, so an utterance naming an NPC is answered in that NPC's Voice and a
-// lone NPC still catches unaddressed speech. The returned Roster is the
-// programmatic control surface for adding/removing NPCs at runtime (#49); the
-// caller owns it for the cycle's lifetime. npcs must be non-empty.
-//
-// language is the Campaign Language of the campaign the npcs belong to: it
-// selects the Roster matcher's phonetic encoder (#199). A code with no
-// registered encoder — including the env-only path's "" — resolves to "en"
-// (see matcherLanguage).
-//
-// All NPCs share ONE tool-engine (one client, the `dice` grant in code — Tool
-// Grants are a #6 table, not yet seeded). The LLM adapter is dispatched off the
-// primary Agent's provider_config.provider via llmProviderID -> [newLLM]
-// ([llmbuild.New], #272): the seed's groq config (model openai/gpt-oss-120b, the
-// #424 default, via the OpenAI-compat endpoint) and the env-only "" both resolve to Groq (ADR-0036),
-// so this stays byte-identical to the pre-#272 hardwired constructor, while a saved
-// non-Groq LLM config now gets its own adapter. Keyless cassette tests replay the
-// Anthropic adapter behind the same llm.Provider interface.
-//
-// synth is the [tts.Synthesizer] the TTS stage drives. [Run] passes a
-// [wire.TeeSynthesizer] wrapping the real ElevenLabs synthesizer so the
-// synthesized audio is tee'd to the playback path while the orchestrator keeps
-// draining-and-dropping it (ADR-0021); a bare ElevenLabs synthesizer also works
-// (no audio is played). It must not be nil.
-// wireMutes subscribes roster.SetMuted to [voiceevent.MuteChanged] on bus (so a
-// GM mute from either surface de-routes / restores the NPC live) and seeds the
-// current mute state from mutes, so a mid-session Discord RECONNECT — which
-// rebuilds the roster from scratch — re-applies the mutes that were in effect
-// (#211). It returns the unsubscribe func; the caller defers it for the cycle's
-// lifetime. A nil mutes view still subscribes (a mute can arrive after connect)
-// but seeds nothing.
-//
-// On each event the roster is set from the AUTHORITATIVE view (mutes.Muted), NOT
-// the event's payload: the Manager can publish two overlapping ops' events in an
-// order that no longer reflects the final set (a mute-all straddling a per-Agent
-// unmute), so trusting a stale payload could leave an NPC de-routed while the
-// Manager says unmuted. Re-reading the view makes every event converge the roster
-// to the current truth — and closes the subscribe-then-seed window too (a mute
-// landing between subscribe and seed applies the same current value either way).
-func wireMutes(bus *voiceevent.Bus, roster *Roster, mutes orchestrator.MuteView) func() {
-	unsub := voiceevent.On(bus, func(e voiceevent.MuteChanged) {
-		muted := e.Muted
-		if mutes != nil {
-			muted = mutes.Muted(e.AgentID) // authoritative re-read; ignore the (possibly stale) payload
-		}
-		roster.SetMuted(e.AgentID, muted)
-	})
-	if mutes != nil {
-		// Seed via the locked reconcile, NOT a raw range over roster.specs: this runs
-		// on connectAndServe's goroutine while the subscription above may fire
-		// SetMuted on the publisher's goroutine (a concurrent GM mute), so the whole
-		// seed must be under the Roster lock.
-		roster.ApplyMutes(mutes.Muted)
-	}
-	return unsub
-}
-
-// tapeInboundOptions returns the [wire.Pipeline] options that copy every consented
-// inbound Speaker frame into the rollover tape (#306). A nil tape (campaign not
-// armed) returns nothing, so the pipeline is byte-identical to the pre-tape loop.
-// The tap runs inline on the audio loop: tape.AppendInbound is non-blocking and
-// drops unconsented Speakers before any buffer (ADR-0051), so it adds no latency.
-func tapeInboundOptions(t *tape.Tape) []wire.Option {
-	if t == nil {
-		return nil
-	}
-	return []wire.Option{
-		wire.WithInboundTap(func(f gxvoice.Frame) {
-			t.AppendInbound(f.UserID.String(), f.Opus, time.Now())
-		}),
-	}
-}
-
-// tapePumpOptions returns the [wire.PlaybackPump] options that copy every agent
-// Opus frame pulled to the wire into the rollover tape's always-on agent lane
-// (#306, ADR-0051). A nil tape returns nothing (unchanged playback).
-func tapePumpOptions(t *tape.Tape) []wire.PumpOption {
-	if t == nil {
-		return nil
-	}
-	return []wire.PumpOption{
-		wire.WithOutboundOpusTap(func(opus []byte) {
-			t.AppendAgent(opus, time.Now())
-		}),
-	}
-}
-
-// buildHighlightDetector constructs the Session Highlights moment detector (#307)
-// for this Voice Session cycle, or returns nil when highlights are off. It is armed
-// ONLY when both the rollover tape (the clip source) and a trigger sink are wired —
-// a detector with nothing to cut clips from, or nowhere to hand triggers, is inert
-// by construction. It reuses the recap/wirenpc BYOK chain (llmbuild.New keyed off
-// cfg.llmProviderID + cfg.keys.llm) for the classifier provider, snapshots via the
-// tape, gates on the session spend gate, and meters on the shared StageRecorder the
-// Manager already tees into the spend meter (ADR-0046). A provider-build failure is
-// logged and degrades to no detector — a Highlight is best-effort and must never
-// fail the session.
-func buildHighlightDetector(cfg Config, bus *voiceevent.Bus, log *slog.Logger) *highlight.Detector {
-	if cfg.Tape == nil || cfg.Highlights == nil {
-		return nil
-	}
-	provider, err := newLLM(cfg.llmProviderID, cfg.keys.llm)
-	if err != nil {
-		log.Warn("highlight detector: build classifier provider; highlights disabled this session", "err", err)
-		return nil
-	}
-	model := ""
-	if len(cfg.npcs) > 0 {
-		model = cfg.npcs[0].model
-	}
-	return highlight.NewDetector(
-		bus,
-		provider,
-		model,
-		cfg.Tape.Snapshot,
-		cfg.Highlights,
-		cfg.Gate,
-		cfg.StageMetrics,
-		log,
-		highlight.Config{ProviderLabel: llmProviderLabel(cfg.llmProviderID)},
-	)
-}
-
-// highlightPCMOptions wires the detector's decoded-PCM tap into the inbound
-// pipeline (#307). A nil detector returns nothing (byte-identical loop). The tap is
-// non-blocking (drops under load), so it adds no audio-loop latency (ADR-0020).
-func highlightPCMOptions(d *highlight.Detector) []wire.Option {
-	if d == nil {
-		return nil
-	}
-	return []wire.Option{wire.WithPCMTap(d.PCMTap())}
-}
-
-// TapeConsentReader is the authoritative consent read the tape reseeds from (#306):
-// the durable tape_consent rows for a Campaign. *storage.Store satisfies it.
-type TapeConsentReader interface {
-	ListTapeConsent(ctx context.Context, campaignID uuid.UUID) ([]string, error)
-}
-
-// wireTapeConsent keeps the tape's consent set converged to the DURABLE truth
-// (#306, ADR-0051), the exact discipline the mute wiring uses (wireMutes): it
-// re-reads the authoritative consent rows rather than trusting an event payload, so
-// two out-of-order events can't leave the tape granted while the DB says revoked.
-//
-//   - It reseeds from ListTapeConsent at cycle start, so a grant/revoke that landed
-//     while nothing was subscribed (a reconnect backoff gap — the subscription is
-//     per cycle but the tape is per session) is applied on the next cycle.
-//   - On every TapeConsentChanged it filters e.CampaignID to THIS session's campaign
-//     — a press against a stale disclosure for another campaign (a reused channel)
-//     must not arm a lane here — then re-reads the store and reconciles the whole set.
-//
-// It returns the unsubscribe func for the caller to defer. A nil tape (campaign not
-// armed) does nothing.
-func wireTapeConsent(ctx context.Context, bus *voiceevent.Bus, t *tape.Tape, campaignID uuid.UUID, store TapeConsentReader, log *slog.Logger) func() {
-	if t == nil {
-		return func() {}
-	}
-	if store == nil {
-		// A wiring bug: the tape is armed but no consent reader was threaded through
-		// (RunFromDB sets both together). Refuse to reconcile against a nil store —
-		// which would panic — and log loudly; the tape keeps its construction-time
-		// seed rather than crashing the session.
-		log.Error("tape: armed but no consent reader wired; live consent changes will not apply", "campaign", campaignID)
-		return func() {}
-	}
-	reconcile := func() {
-		consented, err := store.ListTapeConsent(ctx, campaignID)
-		if err != nil {
-			log.Warn("tape: reconcile consent from store", "campaign", campaignID, "err", err)
-			return
-		}
-		t.ReconcileConsent(consented)
-	}
-	// Subscribe BEFORE the seed reconcile so an event landing in the seed window is
-	// not lost (the wireMutes precedent): worst case is one extra idempotent
-	// reconcile, never a dropped consent change.
-	unsub := voiceevent.On(bus, func(e voiceevent.TapeConsentChanged) {
-		if e.CampaignID != campaignID.String() {
-			return // a consent press for a different campaign — not ours
-		}
-		reconcile() // re-read the durable truth; ignore the (possibly stale/reordered) payload
-	})
-	reconcile() // authoritative reseed at cycle start (catches changes during a reconnect gap)
-	return unsub
-}
-
-// The playback pump is the production [orchestrator.LookaheadPump] (#375): the same
-// serialized playback path holds and releases the Cross-talk Reaction's pre-rendered
-// first sentence.
-var _ orchestrator.LookaheadPump = (*wire.PlaybackPump)(nil)
-
-func buildConversation(bus *voiceevent.Bus, log *slog.Logger, npcs []npcSpec, language string, synth tts.Synthesizer, stageMetrics observe.StageRecorder, keys providerKeys, llmProviderID string, streaming bool, memory agent.MemoryRecaller, facts agent.FactsRecaller, mutes orchestrator.MuteView, gate orchestrator.TurnGate, gmSpeaker func(speakerID string) bool, toolDeps tool.Deps, textPoster func(ctx context.Context, text string) error, lookahead orchestrator.LookaheadPump, clipReplayLoad orchestrator.ClipLoader, clipReplaySink orchestrator.ClipSink) (*orchestrator.Conversation, *Roster, func(), error) {
-	if stageMetrics == nil {
-		stageMetrics = observe.Discard{}
-	}
-	if len(npcs) == 0 {
-		return nil, nil, nil, fmt.Errorf("wirenpc: buildConversation needs at least one NPC")
-	}
-
-	engine, err := silero.New(silero.WithMinSilenceFrames(vadMinSilenceFrames))
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("init Silero VAD: %w", err)
-	}
-	vadSession, err := engine.NewSession(vad.Config{
-		SampleRate:       vadSampleRate,
-		FrameSizeMs:      vadFrameMs,
-		SpeechThreshold:  0.5,
-		SilenceThreshold: 0.35,
-	})
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("open VAD session: %w", err)
-	}
-	// cleanup releases ONLY the per-cycle VAD session (its ONNX inferencer), which
-	// a reconnect cycle (issue #44) would otherwise leak on every loop. It must
-	// NOT close the engine: silero.Engine wraps the process-global ONNX
-	// environment (initialised once via sync.Once and never re-initialised), so
-	// engine.Close() → DestroyEnvironment would tear ONNX down for the whole
-	// process — the next cycle's NewSession would fail and the NPC would go
-	// permanently deaf after the first Discord drop. The engine is a singleton: it
-	// lives for the process and is never closed here. session.Close is idempotent.
-	cleanup := func() {
-		_ = vadSession.Close()
-	}
-	// vad_hangover (#125): the fixed end-of-speech detection lag is a constant the
-	// stage cannot read off the Silero session (vad.Session doesn't expose
-	// minSilenceFrames), so compute it here from the same consts the session is
-	// configured with — vadMinSilenceFrames*vadFrameMs (= 384 ms).
-	vadHangover := vadMinSilenceFrames * vadFrameMs * time.Millisecond
-	vadStage := orchestrator.NewVAD(bus, vadSession,
-		orchestrator.WithVADMetrics(stageMetrics, vadHangover))
-
-	// Speaker Lanes (ADR-0050): each Discord participant's already-separated frame
-	// stream (codec stamps Frame.UserID) opens its own VAD lane, so utterances are
-	// segmented and attributed per speaker. The lane factory builds a fresh Silero
-	// session PER lane from the SAME process-global engine (NewSession is cheap; the
-	// engine is the ONNX singleton, never re-created) and its close releases that
-	// lane's ONNX inferencer on reap (risk (b)). The default lane keeps vadSession;
-	// the factory builds only the non-default lanes.
-	laneVADFactory := func() (*orchestrator.VAD, func(), error) {
-		sess, err := engine.NewSession(vad.Config{
-			SampleRate:       vadSampleRate,
-			FrameSizeMs:      vadFrameMs,
-			SpeechThreshold:  0.5,
-			SilenceThreshold: 0.35,
-		})
-		if err != nil {
-			return nil, nil, fmt.Errorf("open lane VAD session: %w", err)
-		}
-		v := orchestrator.NewVAD(bus, sess, orchestrator.WithVADMetrics(stageMetrics, vadHangover))
-		return v, func() { _ = sess.Close() }, nil
-	}
-
-	// Bounded transient-error retry (#124, ADR-0044): one shared policy threaded
-	// into all three provider stages (STT, TTS, LLM). Log-only per-attempt detail;
-	// the injected time is the production timer (a live run, not a cassette), and the
-	// per-stage deadlines stay the hard bound the loop never extends.
-	retryPolicy := retry.Policy{Log: log}
-
-	// One recognizer instance backs both the batch stage and — when streaming is
-	// enabled and the adapter supports it — the stream manager (the ElevenLabs
-	// Client is both a batch stt.Recognizer and a stt.StreamingRecognizer, ADR-0042).
-	var recognizer stt.Recognizer = newSTT(keys.stt)
-	sttStage := orchestrator.NewSTT(bus, recognizer,
-		orchestrator.WithSTTMetrics(stageMetrics, observe.ProviderElevenLabs),
-		orchestrator.WithSTTRetry(retryPolicy))
-	// tts_total + tts-stage provider health (#125): ElevenLabs is the wired TTS
-	// provider (ADR-0039), so the spans are labelled elevenlabs.
-	ttsStage := orchestrator.NewTTS(bus, synth,
-		orchestrator.WithTTSMetrics(stageMetrics, observe.ProviderElevenLabs),
-		orchestrator.WithTTSRetry(retryPolicy))
-	streamMgr := buildStreamManager(recognizer, streaming, stageMetrics, log)
-	// Per-lane streaming STT (ADR-0042 × ADR-0050): each Speaker Lane opens its own
-	// stream (stamping its SpeakerID on partials) under a concurrency cap, so
-	// concurrent sockets track concurrent speakers, not channel size. Nil unless
-	// streaming is on AND the adapter streams — the byte-for-byte batch default.
-	laneStreamFactory := buildLaneStreamFactory(recognizer, streaming, stageMetrics)
-
-	// Tool Grants are now DB-backed and hydrated per NPC (#113, ADR-0029): each
-	// NPC carries its own grants (spec.grants — from its tool_agent_grant rows on
-	// the DB path, the in-code default on the env-only path), resolved into a
-	// per-NPC GrantSet against ONE shared Registry. So the LLM only ever sees the
-	// Tools that NPC is granted, and an NPC with no grant rows is shown no Tool at
-	// all — least-privilege, data-driven, not compiled in.
-	//
-	// Groq is the live LLM provider (see the function doc). Its key is keys.llm:
-	// the decrypted saved BYOK key (issue #69) when one is configured, otherwise
-	// "" so the adapter falls back to GROQ_API_KEY at request time (BYOK,
-	// ADR-0004) — export it from the keyring before an env-only live run
-	// (docs/agents/live-npc-run.md). There is no Anthropic key, so wiring the
-	// Anthropic adapter here would pass the keyless cassette tests (which replay
-	// Anthropic) but fail the live run — Groq is the only correct default for a
-	// runnable NPC. The Groq client and the Registry are shared across every NPC
-	// — only the GrantSet differs per NPC — so N NPCs reuse one client rather than
-	// each opening their own.
-	// Dispatch the LLM adapter off the Agent's Provider Config provider id (#272):
-	// an empty id (env-only path, or a nil LLM config) resolves to the Groq default
-	// (ADR-0036), so the seed's groq config stays byte-identical and the keyless
-	// cassette gate holds. A saved non-Groq provider now gets its own adapter.
-	provider, err := newLLM(llmProviderID, keys.llm)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("wirenpc: build LLM provider: %w", err)
-	}
-	reg := tool.BuiltinRegistry(toolDeps)
-	engineFor := engineFactory(provider, reg, language, stageMetrics, llmProviderLabel(llmProviderID), retryPolicy)
-
-	// Assemble the initial roster: each AddNPC registers the NPC's routing Agent
-	// in the Matcher and its Replier (over a per-NPC engine carrying that NPC's
-	// GrantSet) in the Cast. The Matcher is built from the first NPC and grown for
-	// the rest.
-	roster := newRoster(rosterDepsForLive(engineFor, newTTS(keys.tts), 16, log, memory, facts, language, gmSpeaker, textPoster))
-	for _, npc := range npcs {
-		roster.AddNPC(npc)
-	}
-
-	// Butler GM-only voice-address gate (#280, ADR-0024): WithButlerGMGate reads
-	// the utterance's SpeakerID (ADR-0050 attribution) to keep the Butler a
-	// GM-only voice address. A nil gmSpeaker (voice standalone / bench) passes a
-	// nil gate — off, so every Butler route publishes exactly as before.
-	detector := newAddressDetector(roster.matcher, orchestrator.WithButlerGMGate(gmSpeaker))
-
-	conv := orchestrator.NewConversation(bus, vadStage, sttStage, ttsStage,
-		orchestrator.WithDetector(detector),
-		// B1: stream the reply sentence-by-sentence so first audio begins after the
-		// first sentence, not the whole completion. The agenttool Engine implements
-		// agent.StreamingEngine (it streams the final answer round), so the Cast's
-		// reply stream dispatches each sentence as it lands, to the addressed NPC.
-		orchestrator.WithReplyStream(roster.cast.ReplyStream()),
-		// Barge-in (ADR-0027): a human talking over Bart cancels his turn. The
-		// confirm window must be > 0 against a live mic — a zero window let the
-		// addressing user's own continued speech (single shared VAD session, no
-		// speaker identity) cancel the turn it had just triggered, which is the 20s
-		// self-cancel the latency investigation found. With B1 a confirmed barge
-		// cancels mid-generation, not just pending dispatch.
-		orchestrator.WithBargeInCoalesce(bargeConfirmWindow, floorCoalesceWindow),
-		// Streaming STT (ADR-0042, issue #180): a nil manager is byte-for-byte the
-		// batch path, so this option is unconditional — buildStreamManager returns nil
-		// unless GLYPHOXA_STT_STREAMING is set AND the adapter is a StreamingRecognizer.
-		orchestrator.WithStreamingSTT(streamMgr),
-		// Speaker Lanes (ADR-0050): attributed inbound frames open per-speaker VAD
-		// lanes so utterances are segmented and attributed independently. A single
-		// participant still gets one lane transcribed exactly as today; the default
-		// (unattributed) lane carries the silence clock. Unconditional — the factory is
-		// only invoked for a non-empty SpeakerID.
-		orchestrator.WithSpeakerLanes(laneVADFactory),
-		// Per-lane streaming STT under the env cap (ADR-0042 × ADR-0050). A nil factory
-		// (batch default) leaves the lanes batch-only, so this is unconditional.
-		orchestrator.WithLaneStreamingSTT(laneStreamFactory, streamMaxLanes()),
-		// Per-Agent mute (#211): the replier discards a muted addressee's route
-		// before taking the floor, and a MuteCut reactor cuts the floor when the
-		// speaking Agent is muted. A nil view is the feature-off default (voice
-		// standalone / bench), so this option is unconditional.
-		orchestrator.WithMute(mutes),
-		// Per-session spend soft cap (#130, ADR-0046): the replier refuses a NEW turn
-		// once the session's estimated spend crosses the soft cap. A nil gate is the
-		// feature-off default (no caps configured), so this option is unconditional.
-		orchestrator.WithTurnGate(gate),
-		// GM /say direct speech (#295, ADR-0010): a DirectSpeech reactor renders a
-		// SpeakRequested (the /say slash command) to TTS in the addressed NPC's Voice,
-		// looked up from THIS roster. It shares the barge-in floor (so a human barge
-		// cancels a /say) and the spend gate, but bypasses mute (GM puppeteering). The
-		// session Manager publishes SpeakRequested; the lookup is always wired so /say
-		// works whenever a session is live.
-		orchestrator.WithDirectSpeech(roster.Voice),
-		// Pump look-ahead for the Cross-talk Reaction onset gap (#375, ADR-0025): the
-		// playback pump doubles as the [orchestrator.LookaheadPump], so a queued
-		// Reaction's first sentence pre-renders its audio during the Lead's playback.
-		// A nil pump is the feature-off default (TEXT-only pre-render); it only takes
-		// effect alongside an Ensemble speaker, so it is safe to wire unconditionally.
-		orchestrator.WithReactionLookahead(lookahead),
-		// Highlight voice replay (#310, ADR-0051): a ClipReplay reactor plays a promoted
-		// Highlight's clip into the live voice channel on a ReplayRequested, loading the
-		// clip via clipReplayLoad and pushing its chunks to the session's PlaybackPump.
-		// It shares the barge-in floor (so a human barge cancels a replay). A nil loader
-		// is the feature-off default — the sink is always the live pump, so this is safe
-		// to wire unconditionally; Register only binds the reactor when the loader is set.
-		orchestrator.WithClipReplay(clipReplayLoad, clipReplaySink),
-		// Handles failures the reactors fire off the audio loop: the replier's TTS
-		// dispatch and the segmenter's off-loop STT call (#24). The wrapped error
-		// names its stage (orchestrator.TTS.Dispatch / orchestrator.STT.Transcribe).
-		orchestrator.WithErrorHandler(func(err error) {
-			log.Warn("voice pipeline stage failed", "err", err)
-		}),
-	)
-	return conv, roster, cleanup, nil
-}
-
-// engineFactory builds the per-NPC engine constructor buildConversation hands the
-// Roster: one shared Groq provider + Registry, one engine per NPC differing only
-// by GrantSet and model. It is a named function (not an inline closure) so the
-// model-threading contract (#227) is unit-testable with a fake provider that
-// captures the [llm.Request].
-//
-// spec.model — resolved from the Agent's LLM provider_config (loadCampaignNPCs),
-// tenant-level row as fallback — flows verbatim into the request. Empty passes
-// through as "" so the openaicompat adapter fills [groq.DefaultModel]; there is
-// NO defaulting duplicated here (the AC "empty → provider default" is proven at
-// the adapter). language selects the dice gate's keyword set (#226). provName is
-// the ACTUAL LLM provider (#272): it labels the per-round LLM spans (A3) AND keys
-// the spend price lookup, so a non-Groq adapter is not mispriced as groq.
-func engineFactory(provider llm.Provider, reg *tool.Registry, language string, stageMetrics observe.StageRecorder, provName observe.Provider, retryPolicy retry.Policy) func(npcSpec) agent.Engine {
-	return func(spec npcSpec) agent.Engine {
-		return agenttool.NewEngine(provider, tool.NewGrantSet(reg, spec.grants...), spec.agentID, spec.model, 0, 0,
-			// The per-round LLM spans (A3) and the spend price lookup are labelled with
-			// the actual provider (#272). The no-op recorder keeps the keyless path
-			// silent; the live binary / benchmark inject a real one.
-			agenttool.WithMetrics(stageMetrics, provName),
-			// The dice gate selects its keyword set by Campaign Language (#226), so a
-			// German campaign arms the dice Tool for German roll requests instead of
-			// gating it out and forcing the model to improvise the roll.
-			agenttool.WithLanguage(language),
-			// Bounded transient-error retry around the LLM start call (#124, ADR-0044):
-			// the same shared policy the STT/TTS stages carry, so one 429/5xx rides
-			// through inside the per-turn deadline.
-			agenttool.WithRetry(retryPolicy))
-	}
-}
-
-// buildStreamManager returns the streaming-STT manager when streaming is enabled
-// AND the wired recognizer implements [stt.StreamingRecognizer], else nil (the
-// byte-for-byte batch default). It is the selection seam (ADR-0042, issue #180):
-// keeping it a small pure function lets the gating be unit-tested without standing
-// up the whole Silero/ONNX pipeline. The provider label is elevenlabs (the only
-// streaming STT adapter in the MVP matrix, ADR-0039).
-func buildStreamManager(recognizer stt.Recognizer, streaming bool, stageMetrics observe.StageRecorder, log *slog.Logger) *orchestrator.StreamManager {
-	if !streaming {
-		return nil
-	}
-	sr, ok := recognizer.(stt.StreamingRecognizer)
-	if !ok {
-		// Opting in but the wired provider can't stream is a misconfiguration worth
-		// surfacing — otherwise the operator sees the batch path with no hint why.
-		if log != nil {
-			log.Warn("STT streaming requested but provider does not support it; using batch")
-		}
-		return nil
-	}
-	return orchestrator.NewStreamManager(sr,
-		orchestrator.WithStreamMetrics(stageMetrics, observe.ProviderElevenLabs))
-}
-
-// defaultStreamMaxLanes caps concurrent per-lane streaming-STT sockets (ADR-0050):
-// concurrent connections equal concurrent speakers, not channel size. Sized for a
-// typical active table; past it a lane transcribes pure batch.
-const defaultStreamMaxLanes = 4
-
-// streamMaxLanes reads the per-lane streaming concurrency cap from
-// GLYPHOXA_STT_STREAM_MAX_LANES. Absent, empty, or unparseable → [defaultStreamMaxLanes]
-// (a typo must not silently change behaviour). An explicit "0" IS honoured — it
-// disables per-lane streaming (0 lanes stream) while the default lane keeps its own
-// stream; a negative value is invalid and falls back to the default (finding 7).
-func streamMaxLanes() int {
-	v := os.Getenv("GLYPHOXA_STT_STREAM_MAX_LANES")
-	if v == "" {
-		return defaultStreamMaxLanes
-	}
-	n, err := strconv.Atoi(v)
-	if err != nil || n < 0 {
-		return defaultStreamMaxLanes
-	}
-	return n // 0 = explicitly disabled
-}
-
-// buildLaneStreamFactory returns the per-Speaker-Lane [orchestrator.StreamManager]
-// factory when streaming is enabled AND the wired recognizer streams, else nil (the
-// batch default). It is the lane analogue of [buildStreamManager]: each lane's
-// manager stamps its SpeakerID on partials (ADR-0050). The provider label is
-// elevenlabs (the only streaming STT adapter in the MVP matrix, ADR-0039).
-func buildLaneStreamFactory(recognizer stt.Recognizer, streaming bool, stageMetrics observe.StageRecorder) func(speakerID string) *orchestrator.StreamManager {
-	if !streaming {
-		return nil
-	}
-	sr, ok := recognizer.(stt.StreamingRecognizer)
-	if !ok {
-		return nil
-	}
-	return func(speakerID string) *orchestrator.StreamManager {
-		return orchestrator.NewStreamManager(sr,
-			orchestrator.WithStreamMetrics(stageMetrics, observe.ProviderElevenLabs),
-			orchestrator.WithStreamSpeakerID(speakerID))
-	}
 }
